# Frontend/src/Adapters/seatpicker.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Infrastructure",
    "version": "1.0"
  },
  "paths": {
    "/seat": {
      "post": {
        "tags": [
          "seat"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Seat.Create.Endpoint (Infrastructure)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeatCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/reservationmanagement": {
      "post": {
        "tags": [
          "reservationmanagement"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.ReservationManagement.Create.Endpoint (Infrastructure)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationManagementCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationManagementCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationManagementCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/reservation/{id}": {
      "put": {
        "tags": [
          "reservation"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Reservation.Move.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationMoveRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationMoveRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationMoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "reservation"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Reservation.Remove.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/reservationmanagement/{id}": {
      "put": {
        "tags": [
          "reservationmanagement"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.ReservationManagement.Move.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationManagementMoveRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationManagementMoveRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationManagementMoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "reservationmanagement"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.ReservationManagement.Remove.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/seat/{id}": {
      "delete": {
        "tags": [
          "seat"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Seat.Remove.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "seat"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Seat.Update.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeatUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Bounds": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "format": "double"
          },
          "y": {
            "type": "number",
            "format": "double"
          },
          "width": {
            "type": "number",
            "format": "double"
          },
          "height": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ReservationManagementCreateRequest": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ReservationManagementMoveRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "fromSeatId": {
            "type": "string",
            "format": "uuid"
          },
          "toSeatId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ReservationMoveRequest": {
        "type": "object",
        "properties": {
          "fromSeatId": {
            "type": "string",
            "format": "uuid"
          },
          "toSeatId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "SeatCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "bounds": {
            "$ref": "#/components/schemas/Bounds"
          }
        },
        "additionalProperties": false
      },
      "SeatUpdateRequest": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "bounds": {
            "$ref": "#/components/schemas/Bounds"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Bearer token acquired through the Discord authentication endpoints",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}